Add tag_list to Tags class so its order is reliable
Before this the list of tags contained on a Tags instance
was stored on a python Counter class, i.e. a python Dict.

As a result saving, and restoring a Tags class on disk
using pickle would cause this order to change and thus
messing the order of the feature vector.

Now we have a list to make sure this order is always kept the same.